Cambio de flotantes a enteros y mejoras en codigo
 Se mejoro la condicion al querer cargar un archivo y no existe

+++ simulacion/Coordenada.cs
@@ -8,25 +8,25 @@ namespace simulacion
 {
     class Coordenada
     {
-        private float x;
-        private float y;
+        private int x;
+        private int y;
 
-        public void setX(float x)
+        public void setX(int x)
         {
             this.x = x;
         }
 
-        public void setY(float y)
+        public void setY(int y)
         {
             this.y = y;
         }
 
-        public float getX()
+        public int getX()
         {
             return x;
         }
 
-        public float getY()
+        public int getY()
         {
             return y;
         }
